fix: address window api review
